adding len(data) condition

+++ math/0x03-probability/binomial.py
@@ -25,8 +25,8 @@ class Binomial:
                     sum_x = sum_x + (data[i] - ux)**2
                 var_x = sum_x / len(data)
                 P = 1 - (var_x / ux)
-                self.n = round(ux / P)
-                self.p = ux / self.n
+                self.n = int(round(ux / P))
+                self.p = float(ux / self.n)
 
     def pmf(self, k):
         """calculates the PMF for a given number of successes: k"""
